chore(RHIF-279): remove circular dependency
