+++ src/components/ImmutableDevices/helpers.js
@@ -1,4 +1,3 @@
-import React from 'react';
 import {
   CheckCircleIcon,
   CubeIcon,
@@ -18,7 +17,6 @@ import warningColor from '@patternfly/react-tokens/dist/esm/global_warning_color
 import successColor from '@patternfly/react-tokens/dist/esm/global_success_color_100';
 import infoColor from '@patternfly/react-tokens/dist/esm/global_info_color_100';
 import activeColor from '@patternfly/react-tokens/dist/esm/global_active_color_100';
-import Status from './Status';
 
 export const colorMapper = {
   green: successColor.value,
@@ -66,50 +64,3 @@ export const getDeviceStatus = (
     return 'upToDate';
   }
 };
-
-export const edgeColumns = [
-  {
-    key: 'ImageName',
-    title: 'Image',
-    sort: false,
-    renderFunc: (imageName, uuid) => {
-      return <a href={`/edge/inventory/${uuid}`}>{imageName}</a>;
-    },
-    props: { isStatic: true },
-  },
-  {
-    key: 'Status',
-    title: 'Status',
-    sort: false,
-    renderFunc: (
-      StatusText,
-      DEVICE_ID,
-      { UpdateAvailable, DispatcherStatus }
-    ) => {
-      const deviceStatus = getDeviceStatus(
-        StatusText,
-        UpdateAvailable,
-        DispatcherStatus
-      );
-
-      return deviceStatus === 'error' || deviceStatus === 'unresponsive' ? (
-        <Status
-          type={
-            deviceStatus === 'error'
-              ? 'errorWithExclamationCircle'
-              : deviceStatus
-          }
-          isLink={true}
-        />
-      ) : (
-        <Status
-          type={
-            deviceStatus === 'error'
-              ? 'errorWithExclamationCircle'
-              : deviceStatus
-          }
-        />
-      );
-    },
-  },
-];
