+++ src/components/ImmutableDevices/index.js
@@ -1,0 +1,1 @@
+export { default as ImmutableDevices } from './ImmutableDevices';
